front-end/requirements.txt
google-cloud-storage==2.9.0
google-cloud-aiplatform
google-cloud-vision
streamlit==1.31.1
st-pages==0.4.5
vertexai==1.70.0
google-generativeai==0.8.3
google-cloud-texttospeech==2.21.0
